soulspark/backend/requirements.txt
fastapi
uvicorn[standard]
httpx
sqlmodel
pydantic
apscheduler
python-dotenv
google-generativeai
psycopg[binary]
google-auth
PyJWT
